+++ CHANGELOG.md
@@ -2,7 +2,8 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* Upgrade to Pulumi v3.12.0 for sdk and pkg
+* Add support for exporting provider example conversion metrics
 
 ---